#18 Working messaging sockets
Made a helper function for messaging sockets that gets all messages between two users.

Also added conditionals to the sockets to make sure that all the information that goes throught the socket is needed such as the user and userclicked ids. I also added a leave message room socket that on the hit of the back button on a conversation it leaves their room.

Since there was a lot more code added I also made helpful comments that makes the code easier to read.