Routes added for fetching all of a users calendar Items
The way calendarItems were being fetched before was from calendaritems that the user owns and not including the the events that they are attendeeing(added to their calendar)

Added route that creates a calendarItem with userId if the calendar Item they are trying to add is just personal, or if it is an event, then it will create an attendee record that connects the event to the user.

Also made a different delete route for attendees just in case someone who s an attendee of an event and they want to delete the calendar Item off of their calendar, it will delete the attendee record.